refactor(item): extract nested endpoints into separate controller classes
